requirements.txt
psycopg2==2.4.5
gunicorn==0.14.5
https://github.com/django/django/archive/1.5b1.zip
django-geojson-tiles==0.0.4